#8157: [WIP] handle draftjs field
