+++ src/contentScript/textEditorDom.ts
@@ -19,8 +19,21 @@ import * as ckeditorDom from "@/contrib/ckeditor/ckeditorDom";
 import * as pageScript from "@/pageScript/messenger/api";
 import { getSelectorForElement } from "@/contentScript/elementReference";
 import { expectContext } from "@/utils/expectContext";
-import { dispatchPasteForDraftJs } from "@/utils/domFieldUtils";
 import focusController from "@/utils/focusController";
+import {
+  dispatchPasteForDraftJs,
+  isDraftJsField,
+} from "@/contrib/draftjs/draftJsDom";
+
+/**
+ * @file Text Editor DOM utilities that might call the pageScript.
+ *
+ * Historically, we've preferred to use `contentScript` except for calls that must be made from the `pageScript`. The
+ * advantage is that 1) we don't have pageScript injection cold-start, and 2) there's less of a chance of the host page
+ * interfering with our calls.
+ *
+ * However, we might consider consolidating the logic in the pageScript to simplify calling conventions.
+ */
 
 /**
  * Error to be thrown when document.execCommand fails.
@@ -40,21 +53,12 @@ export class ExecCommandError extends Error {
 }
 
 /**
- * @file Text Editor DOM utilities that might call the pageScript.
- *
- * Historically, we've preferred to use `contentScript` except for calls that must be made from the `pageScript`. The
- * advantage is that 1) we don't have pageScript injection coldstart, and 2) there's less of a chance of the host page
- * interfering with our calls.
- *
- * However, we might consider consolidating the logic in the pageScript to simplify calling conventions.
- */
-
-/**
  * Inserts text at the current cursor position in the given element, with support for custom editors, e.g., CKEditor.
  *
  * Current support:
  * - Plain content editable (Gmail, etc.)
  * - CKEditor 4/5
+ * - Draft.js
  *
  * @throws {ExecCommandError} if the text could not be inserted with InsertTextError
  */
@@ -76,9 +80,11 @@ export async function insertAtCursorWithCustomEditorSupport(text: string) {
     return;
   }
 
-  if (element.textContent === "" && element.closest(".DraftEditor-root")) {
-    // Special handling for DraftJS if the field is empty
-    // https://github.com/pixiebrix/pixiebrix-extension/issues/7630
+  if (isDraftJsField(element)) {
+    // Must always use paste. Otherwise, the editor will become corrupted
+    // More information/context:
+    // - https://github.com/pixiebrix/pixiebrix-extension/issues/7630
+    // - https://github.com/pixiebrix/pixiebrix-extension/issues/8157
     dispatchPasteForDraftJs(element, text);
     return;
   }
